// src/types/supplier.ts
import { AutoGeneratedField, SearchParams, SimpleResponse, PaginatedResponse } from './common';

export type SupplierFormData = {
  name: string;
  phone: string;
  email: string;
  address: string;
  taxCode: string;
  website?: string;
  contactPerson: string;
  contactPhone: string;
  contactEmail: string;
  supplierType: string;
  rating: number;
  costSpend?: number;
  costRemain?: number;
  region?: string;
  province?: string;
  district?: string;
};

// export type Supplier = AutoGeneratedField & SupplierFormData;

// Supplier response types sử dụng Supplier entity
export type SupplierSimpleResponse = SimpleResponse<Supplier>;
export type SupplierPaginatedResponse = PaginatedResponse<Supplier>;

export type SupplierParams = SearchParams<SupplierFormData> & {
  supplierType?: string;
  contactPerson?: string;
  rating?: number;
};

export type TSupplier = Supplier;


export type Bank = {
  id: number;
  createdAt: string; // ISO date string
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate: string;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  name: string;
  fullName: string;
  swiftCode: string;
  bankCode: string;
  branchName: string;
  branchCode: string;
  address: string;
  phone: string;
  email: string;
  website: string;
  country: string;
  city: string;
  currency: string;
  isDefault: number;
  refType: string;
  refId: number;
}

export type Supplier = {
  id: number;
  createdAt: string;
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate: string;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  name: string;
  phone: string;
  email: string;
  address: string;
  taxCode: string;
  website: string;
  contactPerson: string;
  contactPhone: string;
  contactEmail: string;
  supplierType: string;
  rating: number;
  costSpend: number;
  costRemain: number;
  region: string;
  province: string;
  district: string;
  banks: Bank[];
}

export type SupplierResponse = {
  data: Supplier;
};